salir

+++ funciones.php
@@ -9,6 +9,7 @@ function viewTeachers($consulta){//de aqui hasta funcion form create teachers qu
             echo"<table>";
             echo "<tr>";
             echo"<th>id</th>";
+            echo"<th>dni</th>";
             echo"<th>name</th>";
             echo "<th>surname</th>";
             echo "<th>title</th>";
@@ -18,13 +19,13 @@ function viewTeachers($consulta){//de aqui hasta funcion form create teachers qu
             echo "</tr>";
             for ($i =0 ; $i < $numlinies ; $i++){
                 $linea = mysqli_fetch_array($consulta);
-            
                 echo "<tr>";
                 echo "<td>".$linea[0]."</td>";
                 echo "<td>".$linea[1]."</td>";
                 echo "<td>".$linea[2]."</td>";
                 echo "<td>".$linea[4]."</td>";
                 echo "<td>".$linea[5]."</td>";
+                echo "<td>".$linea[6]."</td>";
                 echo "<td><a href='modify-teachers.php?num= ".$linea[0]."'>Modify</a></td>";
                 echo "<td><a href='delete-teachers.php?num= ".$linea[0]."'>Delete</a></td>";
                 echo"</tr>";
@@ -132,9 +133,15 @@ function viewCourses($consulta){
         echo "<td>".$linea[4]."</td>";
         echo "<td>".$linea[5]."</td>";
         echo "<td>".$linea[6]."</td>";
+        if($_SESSION['type']=='Admin'){
+            echo "<td><a href='modify-courses.php?num= ".$linea[0]."'>Modify</a></td>";
+            echo "<td><a href='delete-course.php?num= ".$linea[0]."'>Delete</a></td>";
+        }else if($_SESSION['type']=='Teacher'){
+            echo "<td><a href='put-mark.php?num= ".$linea[0]."'>Asignar nota</a></td>";
+        }else if($_SESSION['type']=='Student'){
+            echo "<td><a href='inscribir-curso.php?num= ".$linea[0]."'>Inscribirme</a></td>";
+        }
 
-        echo "<td><a href='modify-courses.php?num= ".$linea[0]."'>Modify</a></td>";
-        echo "<td><a href='delete-course.php?num= ".$linea[0]."'>Delete</a></td>";
         echo"</tr>";
                         
      }
